# perf: reuse URL for each request