@littledivy 2022-07-22
This will be optimized in Deno itself, so that the workaround is no longer necessary